Removed all instances of floats, sprintfs, and printfs in an attempt to reduce program memory space. It dropped by about 35%

+++ lora.h
@@ -35,7 +35,8 @@
 #define LORA_UNLISTED_FAILURE                        INT8_C(-99)
 
 void lora_init(void);
-int8_t lora_send(uint8_t address, double metrics[], uint8_t size);
+
+int8_t lora_send(uint8_t address, int32_t metrics[], uint8_t size);
 bool lora_enable(void);
 bool lora_disable(void);
